0412 플레이어 중간저장
기력 시스템 및 플레이어 마우스 이동 클릭 유지

+++ Assets/Script/UI/InGameUI/HpAndEnergy.cs
@@ -17,7 +17,7 @@ public class HpAndEnergy : MonoBehaviour
     }
     public void EnergyGageTrigger(int ReferenceValue, int Value)
     {
-        this.Energybar.fillAmount = Value / ReferenceValue;
+        this.Energybar.fillAmount = (float)Value / (float)ReferenceValue;
     }
     
 }
